requirements.txt
mysql-connector-python==8.0.18
boto3==1.10.14
